% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_hierarchy_2018}
\alias{area_hierarchy_2018}
\title{Boundaries for Statistical Area 1 2018}
\format{
A data frame with 29889 rows and 8 variables:
\describe{
\item{SA1_2018_NAME}{Name of the SA1}
\item{SA2_2018_CODE}{Code of the SA2}
\item{SA2_2018_NAME}{Name of the SA2}
\item{TA_2018_CODE}{Code of the TA}
\item{TA_2018_NAME}{Name of the TA}
\item{RC_2018_CODE}{Code of the RC}
\item{RC_2018_NAME}{Name of the RC}
\item{LANDWATER_NAME}{Factor variable indicating what land type the SA1 area is}
}
}
\source{
\url{https://datafinder.stats.govt.nz/}
}
\usage{
area_hierarchy_2018
}
\description{
A dataset that classifies statistical areas 1s into larger geographic areas,
such as statistical area 2s, territorial authorities, and regional councils.
}
\keyword{datasets}
