% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_areas.R
\name{get_areas}
\alias{get_areas}
\title{Function to access geographic areas from the db.geonz package}
\usage{
get_areas(area_type, clipped = TRUE)
}
\arguments{
\item{area_type}{A string of the geographic area type to be selected. Must be one of sa1, sa2, ward, ta, rc}

\item{clipped}{if TRUE, returns the geographic areas clipped to the New Zealand Coastline}
}
\value{
A sf dataframe with three columns: the area code, the area name and the area geometry
}
\description{
Function to access geographic areas from the db.geonz package
}
