% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rc_geoms_2018_clipped}
\alias{rc_geoms_2018_clipped}
\title{Boundaries for Regional Councils 2018 (Clipped)}
\format{
A data frame with 17 rows and 3 variables:
\describe{
\item{code}{Code of the RC}
\item{name}{Name of the RC}
\item{geom}{multipolygon of the area boundaries, as lat/long in WGS84}
}
}
\source{
\url{https://datafinder.stats.govt.nz/}
}
\usage{
rc_geoms_2018_clipped
}
\description{
A dataset of regional council (RC) boundaries at 1 January 2018 as defined
by Stats NZ, clipped to the coastline. This clipped version is created for
map creation/cartographic purposes and so does not fully represent the
official full extent boundaries.
}
\keyword{datasets}
