% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa1_geoms_2018}
\alias{sa1_geoms_2018}
\title{Boundaries for Statistical Area 1 2018}
\format{
A data frame with 29889 rows and 2 variables:
\describe{
\item{name}{Name of the SA1}
\item{geom}{multipolygon of the area boundaries, as lat/long in WGS84}
}
}
\source{
\url{https://datafinder.stats.govt.nz/}
}
\usage{
sa1_geoms_2018
}
\description{
A dataset of statistical area 1 (SA1) boundaries at 1 January 2018 as defined by Stats NZ.
}
\keyword{datasets}
