% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_geoms_2018_clipped}
\alias{sa2_geoms_2018_clipped}
\title{Boundaries for Statistical Area 2 2018 (Clipped)}
\format{
A data frame with 2171 rows and 3 variables:
\describe{
\item{code}{Code of the SA2}
\item{name}{Name of the SA2}
\item{geom}{multipolygon of the area boundaries, as lat/long in WGS84}
}
}
\source{
\url{https://datafinder.stats.govt.nz/}
}
\usage{
sa2_geoms_2018_clipped
}
\description{
A dataset of statistical area 2 (SA2) boundaries at 1 January 2018 as defined
by Stats NZ, clipped to the coastline. This clipped version is created for
map creation/cartographic purposes and so does not fully represent the
official full extent boundaries.
}
\keyword{datasets}
