% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ta_geoms_2018_clipped}
\alias{ta_geoms_2018_clipped}
\title{Boundaries for Territorial Authorities 2018 (Clipped)}
\format{
A data frame with 68 rows and 3 variables:
\describe{
\item{code}{Code of the TA}
\item{name}{Name of the TA}
\item{geom}{multipolygon of the area boundaries, as lat/long in WGS84}
}
}
\source{
\url{https://datafinder.stats.govt.nz/}
}
\usage{
ta_geoms_2018_clipped
}
\description{
A dataset of territorial authority (TA) boundaries at 1 January 2018 as defined
by Stats NZ, clipped to the coastline. This clipped version is created for
map creation/cartographic purposes and so does not fully represent the
official full extent boundaries.
}
\keyword{datasets}
